feat: configuring auth

+++ src/main/java/br/gov/quixada/esporte/users/auth/AuthenticationRepository.java
@@ -4,7 +4,7 @@ import br.gov.quixada.esporte.users.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 
-public interface AuthenticationRepository extends JpaRepository<User, Integer> {
+public interface AuthenticationRepository extends JpaRepository<User, Long> {
 
     UserDetails findByUsername(String username);
 }
